pyproject.toml
[project]
name = "mahabharat-agents"
version = "0.1.0"
description = "One person. One AI company. Infinite leverage. Deploy FAANG-level AI agents named after Mahabharata characters to run your software company."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Sai Krishna Madavarapu", email = "msaikrishna399@gmail.com"}
]
keywords = [
    "ai-agents",
    "multi-agent",
    "claude-code",
    "mahabharata",
    "software-company",
    "orchestration",
    "llm-agents",
    "solo-builder"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/thisizmsk-png/mahabharat-agents"
Repository = "https://github.com/thisizmsk-png/mahabharat-agents"
Issues = "https://github.com/thisizmsk-png/mahabharat-agents/issues"

[project.scripts]
mahabharat-install = "mahabharat_agents.cli:install"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mahabharat_agents"]

[tool.hatch.build.targets.wheel.force-include]
"agents" = "mahabharat_agents/data/agents"
"skills" = "mahabharat_agents/data/skills"
"orchestration" = "mahabharat_agents/data/orchestration"
"context" = "mahabharat_agents/data/context"
